--- src/main/java/blackjack/controller/DeckGenerator.java
@@ -1,7 +1,0 @@
-package blackjack.controller;
-
-import blackjack.domain.Deck;
-
-public interface DeckGenerator {
-    Deck generate();
-}
