+++ src/main/java/blackjack/domain/card/Cards.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
+import blackjack.domain.Score;
 import java.util.Collections;
 import java.util.List;
