+++ src/main/java/blackjack/domain/card/DeckGenerator.java
@@ -1,0 +1,5 @@
+package blackjack.domain.card;
+
+public interface DeckGenerator {
+    Deck generate();
+}
