+++ src/main/java/blackjack/domain/card/ShuffledDeckGenerator.java
@@ -1,11 +1,7 @@
-package blackjack.controller;
+package blackjack.domain.card;
 
 import static java.util.stream.Collectors.toList;
 
-import blackjack.domain.Card;
-import blackjack.domain.Deck;
-import blackjack.domain.Rank;
-import blackjack.domain.Suit;
 import java.util.ArrayDeque;
 import java.util.Collections;
 import java.util.List;
