+++ src/test/java/blackjack/domain/card/DeckTest.java
@@ -1,9 +1,9 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalStateException;
 
-import blackjack.controller.MockDeckGenerator;
+import blackjack.domain.MockDeckGenerator;
 import java.util.Collections;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
