+++ src/test/java/blackjack/domain/participant/NameTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
