+++ src/test/java/blackjack/domain/participant/PlayerTest.java
@@ -1,8 +1,12 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
+import blackjack.domain.Bet;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Rank;
+import blackjack.domain.card.Suit;
 import java.util.Collections;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
